chore: refactor uses in page header
